projects/01_fyyur/starter_code/requirements.txt
babel
python-dateutil==2.6.0
flask-moment
flask-wtf
flask_sqlalchemy
flask_migrate